Created a companion application to interact with the driver.
Set up some basic back and forth communication between the driver and the companion app. At this point the driver will report groups of keypresses that had both a keydown and keyup event. Ex, if "a", "b", and "c" are held down and released at the same time the driver will report that information in a single event.