Added a key combination to register the keyboard.
Registered keyboards now have their input blocked from the OS. Ie pressing "a" would notify the companion app but it would not type an "a" into an input field.

Added support for multiple read modes to the keyboard interface. This determines which data the requester will recieve.

Added several functions to simplify registering, unregistering, and otherwise interacting with the keyboard interface.

Added some extra options to the companion app and improved how requests were being sent to the driver.